pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "django-tasks-scheduler"
packages = [
    { include = "scheduler" },
]
version = "1.0.0"
description = "An async job scheduler for django using redis"
readme = "README.md"
keywords = ["redis", "django", "background-jobs", "job-queue", "task-queue", "redis-queue", "scheduled-jobs"]
authors = [
    "Daniel Moran <daniel@moransoftware.ca>",
]
maintainers = [
    "Daniel Moran <daniel@moransoftware.ca>",
]
license = "MIT"
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Environment :: Web Environment',
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Framework :: Django',
    'Framework :: Django :: 4',
    'Framework :: Django :: 4.0',
    'Framework :: Django :: 4.1',
    'Framework :: Django :: 4.2',
    'Framework :: Django :: 3',
    'Framework :: Django :: 3.2',
]
homepage = "https://github.com/dsoftwareinc/django-redis-scheduler"
documentation = "https://django-tasks-scheduler.readthedocs.io/en/latest/"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/dsoftwareinc/django-tasks-scheduler/issues"
"Funding" = "https://github.com/sponsors/cunla"

[tool.poetry.dependencies]
python = "^3.9"
django = ">=3.2"
croniter = "^1.4"
click = "^8.1"
rq = "^1.15"
pyyaml = { version = "^6.0", optional = true }

[tool.poetry.dev-dependencies]
poetry = "^1.5"
coverage = "^7"
fakeredis = { version = "^2.15", extras = ['lua'] }
Flake8-pyproject = "^1.2"

[tool.poetry.extras]
yaml = ["pyyaml"]

[tool.flake8]
max-line-length = 119
exclude = [
    'scheduler/migrations',
]